Move launcher cases to common
Remove duplicate cases

[stage-19]
