+++ test/e2e/apps.scenarios.js
@@ -1,4 +1,3 @@
-require('./launcher/launcher.cases.js');
 require('./schedules/schedules.cases.js');
 require('./displays/displays.cases.js');
 require('./editor/editor.cases.js');
