+++ test/e2e/common-header/cases/company-subcompanies.js
@@ -12,7 +12,7 @@
   var CompanySettingsModalPage = require('./../pages/companySettingsModalPage.js');
   var MoveCompanyModalPage = require('./../pages/moveCompanyModalPage.js');
   var SafeDeleteModalPage = require('./../pages/safeDeleteModalPage.js');
-  var SignInPage = require('./../../launcher/pages/signInPage.js');
+  var SignInPage = require('./../../common/pages/signInPage.js');
 
   var CompanySubcompaniesScenarios = function() {
