+++ test/e2e/common-header/cases/user-settings.js
@@ -8,7 +8,7 @@
   var CommonHeaderPage = require('./../pages/commonHeaderPage.js');
   var HomePage = require('./../pages/homepage.js');
   var UserSettingsModalPage = require('./../pages/userSettingsModalPage.js');
-  var SignInPage = require('./../../launcher/pages/signInPage.js');
+  var SignInPage = require('./../../common/pages/signInPage.js');
 
   var UserSettingsScenarios = function() {
