+++ test/e2e/common/cases/homepage.js
