+++ test/e2e/common/cases/signin-custom.js
