+++ test/e2e/common/cases/weekly-templates.js
