+++ test/e2e/common/pages/homepage.js
