+++ test/e2e/common/pages/signInPage.js
