+++ test/e2e/common/pages/signUpPage.js
