+++ test/e2e/common/pages/weeklyTemplates.js
