+++ test/e2e/storage/cases/dismiss-modal.js
@@ -3,7 +3,7 @@ var expect = require('rv-common-e2e').expect;
 var helper = require('rv-common-e2e').helper;
 var StorageSelectorModalPage = require('./../pages/storageSelectorModalPage.js');
 var NewFolderModalPage = require('./../pages/newFolderModalPage.js');
-var HomePage = require('./../../launcher/pages/homepage.js');
+var HomePage = require('./../../common/pages/homepage.js');
 var CommonHeaderPage = require('./../../common-header/pages/commonHeaderPage.js');
 var StorageHelper = require('./../pages/helper.js');
 var StorageHomePage = require('./../pages/storageHomePage.js');
